Implement login functionality
Login page is made with passport and passport-local. Login page now redirects to /dashboard when logging in with proper credentials and redirects back to /users/login when logging in with bad credentials. /dashboard can still be accessed even without logging in but it would cause an error due to the view accessing a property in the request body which results to undefined.